study-tools/writing-practice/requirements.txt
streamlit
Pillow
manga-ocr
openai
python-dotenv
requests 